DataDisplay

+++ MVVM/Model/BookingModel.cs
@@ -7,7 +7,8 @@
         public string AccommodationType { get; set; }
         public DateTime? BookingDate { get; set; }
         public int NumberOfNights { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         // Other properties as needed
     }
 }
+
